// apps/backend/src/model/kanel/public/UserOrganization.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { UserId } from './User';
import type { OrganizationId } from './Organization';

/** Identifier type for public.User_Organization */
export type UserOrganizationId = number & { __brand: 'public.User_Organization' };

/** Represents the table public.User_Organization */
export default interface UserOrganization {
  id: UserOrganizationId;

  user_id: UserId | null;

  organization_id: OrganizationId | null;
}

/** Represents the initializer for the table public.User_Organization */
export interface UserOrganizationInitializer {
  /** Default value: nextval('"User_Organization_id_seq"'::regclass) */
  id?: UserOrganizationId;

  user_id?: UserId | null;

  organization_id?: OrganizationId | null;
}

/** Represents the mutator for the table public.User_Organization */
export interface UserOrganizationMutator {
  id?: UserOrganizationId;

  user_id?: UserId | null;

  organization_id?: OrganizationId | null;
}